--- test/e2e/fixtures/docsify-init.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
-    <meta name="viewport" content="width=device-width,initial-scale=1" />
-    <meta charset="UTF-8" />
-    <title>docsify</title>
-  </head>
-  <body>
-    <div id="app"></div>
-  </body>
-</html>
