--- test/e2e/fixtures/html5.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-  </head>
-  <body></body>
-</html>
