+++ test/fixtures/index.html
@@ -1,0 +1,1 @@
+<!-- Blank page for test server -->
